Updated to 3.600+

+++ Items/Weapons/Technical/Novite.cs
@@ -14,24 +14,25 @@ using SGAmod.Projectiles;
 
 namespace SGAmod.Items.Weapons.Technical
 {
-	public class NoviteKnife : SeriousSamWeapon
+	public class NoviteKnife : SeriousSamWeapon, ITechItem
 	{
 		public override void SetStaticDefaults()
 		{
 			DisplayName.SetDefault("Novite Knife");
-			Tooltip.SetDefault("Instantly hits against targets where you swing\nHitting some types of targets in the rear will backstab, automatically becoming a crit");
+			Tooltip.SetDefault("Instantly hits against targets where you swing\nHitting some types of targets in the rear will backstab, automatically becoming a crit\nHolding this weapon increases your movement speed and jump height");
 		}
 		public override void SetDefaults()
 		{
 			base.SetDefaults();
 
-			item.damage = 15;
+			item.damage = 36;
+			item.crit = 25;
 			item.width = 48;
 			item.height = 48;
 			item.melee = true;
 			item.useTurn = true;
-			item.rare = 1;
-			item.value = 2000;
+			item.rare = ItemRarityID.Green;
+			item.value = 2500;
 			item.useStyle = 1;
 			item.useAnimation = 50;
 			item.useTime = 50;
@@ -65,19 +66,19 @@ namespace SGAmod.Items.Weapons.Technical
 
 	}
 
-	public class NoviteStab : ModProjectile
+	public class NoviteStab : ModProjectile,ITrueMeleeProjectile
 	{
 		public override void SetDefaults()
 		{
-			projectile.width = 12;
-			projectile.height = 12;
+			projectile.width = 16;
+			projectile.height = 16;
 			projectile.aiStyle = -1;
 			projectile.friendly = true;
 			projectile.hostile = false;
 			projectile.penetrate = 1;
 			projectile.melee = true;
-			projectile.timeLeft = 30;
-			projectile.extraUpdates = 30;
+			projectile.timeLeft = 40;
+			projectile.extraUpdates = 40;
 			aiType = -1;
 			Main.projFrames[projectile.type] = 1;
 		}
@@ -115,13 +116,13 @@ namespace SGAmod.Items.Weapons.Technical
 		}
 	}
 
-	public class NoviteBlaster : SeriousSamWeapon
+	public class NoviteBlaster : SeriousSamWeapon, ITechItem
 	{
 		private bool altfired = false;
 		public override void SetStaticDefaults()
 		{
 			DisplayName.SetDefault("Novite Blaster");
-			Tooltip.SetDefault("Firing a piercing bolt of electricity\nConsumes Electric Charge, 50 up front, 200 over time to charge up\nHold the fire button to charge a stronger, more accurate shot\nCan deal up to 3X damage and chain once at max charge");
+			Tooltip.SetDefault("Fires a piercing bolt of electricity\nConsumes Electric Charge, 50 up front, 200 over time to charge up\nHold the fire button to charge a stronger, more accurate shot\nCan deal up to 3X damage and chain once at max charge");
 		}
 
 		public override void SetDefaults()
@@ -134,7 +135,7 @@ namespace SGAmod.Items.Weapons.Technical
 			item.useAnimation = 15;
 			item.useStyle = 5;
 			item.noMelee = true;
-			item.knockBack = 2;
+			item.knockBack = 0;
 			item.value = Item.buyPrice(0, 0, 25, 0);
 			item.rare = 2;
 			//item.UseSound = SoundID.Item99;
@@ -148,9 +149,7 @@ namespace SGAmod.Items.Weapons.Technical
 
 		public override bool CanUseItem(Player player)
 		{
-			if (player.SGAPly().electricCharge > 50)
-				return true;
-			return false;
+			return player.SGAPly().ConsumeElectricCharge(50, 0, consume: false);
 		}
 		public override void AddRecipes()
 		{
@@ -165,9 +164,9 @@ namespace SGAmod.Items.Weapons.Technical
 		{
 			float rotation = MathHelper.ToRadians(0);
 			position += Vector2.Normalize(new Vector2(speedX, speedY)) * 8f;
-			if (player.ownedProjectileCounts[mod.ProjectileType("NovaBlasterCharging")] < 1)
+			if (player.ownedProjectileCounts[ModContent.ProjectileType<NovaBlasterCharging>()] < 1)
 			{
-				int proj = Projectile.NewProjectile(position.X, position.Y, speedX, speedY, mod.ProjectileType("NovaBlasterCharging"), damage, knockBack, player.whoAmI);
+				int proj = Projectile.NewProjectile(position.X, position.Y, speedX, speedY, ModContent.ProjectileType<NovaBlasterCharging>(), damage, knockBack, player.whoAmI);
 				player.SGAPly().ConsumeElectricCharge(50, 100);
 			}
 			return false;
@@ -182,6 +181,10 @@ namespace SGAmod.Items.Weapons.Technical
 		public virtual float velocity => 32f;
 		public virtual float spacing => 24f;
 		public virtual int fireRate => 5;
+		public virtual int FireCount => 1;
+		//public virtual float ForcedLock => 1f;
+		public virtual (float,float) AimSpeed => (1f,0f);
+		public int firedCount = 0;
 		protected Player player;
 		public override void SetStaticDefaults()
 		{
@@ -282,7 +285,7 @@ namespace SGAmod.Items.Weapons.Technical
 
 			projectile.Center += projectile.velocity;
 
-			int prog = Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, perturbedSpeed.X, perturbedSpeed.Y, ModContent.ProjectileType<CBreakerBolt>(), (int)((float)projectile.damage * (1f + (perc * 2f))), 3f, player.whoAmI, perc >= 0.99f ? 1 : 0, 0.50f + (perc * 0.20f));
+			int prog = Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, perturbedSpeed.X, perturbedSpeed.Y, ModContent.ProjectileType<CBreakerBolt>(), (int)((float)projectile.damage * (1f + (perc * 2f))), projectile.knockBack, player.whoAmI, perc >= 0.99f ? 1 : 0, 0.50f + (perc * 0.20f));
 			Main.projectile[prog].localAI[0] = (perc * 0.90f);
 			Main.projectile[prog].magic = true;
 			Main.projectile[prog].melee = false;
@@ -291,6 +294,7 @@ namespace SGAmod.Items.Weapons.Technical
 			IdgProjectile.Sync(prog);
 			Main.PlaySound(SoundID.Item91, player.Center);
 
+			if (firedCount>=FireCount)
 			projectile.Kill();
 		}
 
@@ -308,54 +312,70 @@ namespace SGAmod.Items.Weapons.Technical
 			if (player.dead)
 				projectile.Kill();
 			projectile.timeLeft = 2;
-			player.itemTime = 6;
-			player.itemAnimation = 6;
+
+			/*if (firedCount < FireCount && channeling)
+			{
+				player.itemTime = 6;
+				player.itemAnimation = 6;
+			}*/
+
 			Vector2 direction = (Main.MouseWorld - player.MountedCenter);
 			Vector2 directionmeasure = direction;
 			direction.Normalize();
 
-			if (projectile.ai[0] < chargeuptime + 1)
+			bool cantchargeup = false;
+
+			if (projectile.ai[0] < chargeuptime + 1 && firedCount<1)
 			{
 				if (DoChargeUp())
 					projectile.ai[0] += 1;
+				else
+					cantchargeup = true;
 			}
 
+			bool channeling = ((player.channel || (projectile.ai[0] < 5 && !cantchargeup)) && !player.noItems && !player.CCed);
+			bool aiming = true;// firedCount < FireCount;
 
-			bool channeling = ((player.channel || projectile.ai[0] < 5) && !player.noItems && !player.CCed);
-			projectile.Center = player.MountedCenter + direction * spacing;
-
-			Vector2 mousePos = Main.MouseWorld;
-			if (channeling)
+			if (aiming || channeling)
 			{
+				Vector2 mousePos = Main.MouseWorld;
 				if (projectile.owner == Main.myPlayer)
 				{
 					Vector2 diff = mousePos - player.MountedCenter;
 					diff.Normalize();
-					projectile.velocity = diff;
+					projectile.velocity = Vector2.Lerp(Vector2.Normalize(projectile.velocity),diff, channeling ? AimSpeed.Item1 : AimSpeed.Item2);
 					projectile.direction = Main.MouseWorld.X > player.position.X ? 1 : -1;
 					projectile.netUpdate = true;
 					projectile.Center = mousePos;
 				}
 				int dir = projectile.direction;
 				player.ChangeDir(dir);
-				player.itemTime = fireRate;
-				player.itemAnimation = fireRate;
 
 				player.itemRotation = (float)Math.Atan2(projectile.velocity.Y * dir, projectile.velocity.X * dir);
 				projectile.Center = player.MountedCenter + projectile.velocity * velocity;
+
+				if (channeling)
+				{
+					player.itemTime = fireRate;
+					player.itemAnimation = fireRate;
+				}
 			}
 
+			projectile.Center = player.MountedCenter + Vector2.Normalize(projectile.velocity) * spacing;
 
 			if (projectile.ai[0] > 10)
 			{
 
 				ChargeUpEffects();
 
-				if (!channeling)
+
+
+				if (!channeling && player.itemTime<fireRate && firedCount < FireCount)
 				{
-					player.itemTime = fireRate;
-					player.itemAnimation = fireRate;
-					FireWeapon(direction);
+					firedCount += 1;
+					player.itemTime = fireRate*(firedCount< FireCount ? 2 : 1);
+					player.itemAnimation = fireRate * (firedCount < FireCount ? 2 : 1);
+					FireWeapon(Vector2.Normalize(projectile.velocity));
 				}
 
 			}
@@ -363,7 +383,7 @@ namespace SGAmod.Items.Weapons.Technical
 
 	}
 
-	public class NoviteTowerSummon : SeriousSamWeapon
+	public class NoviteTowerSummon : SeriousSamWeapon, ITechItem
 	{
 		public override void SetStaticDefaults()
 		{
@@ -383,7 +403,7 @@ namespace SGAmod.Items.Weapons.Technical
 			item.useStyle = 1;
 			item.noMelee = true;
 			item.knockBack = 2f;
-			item.value = Item.buyPrice(0, 0, 75, 0);
+			item.value = Item.buyPrice(0, 0, 25, 0);
 			item.rare = 1;
 			item.autoReuse = false;
 			item.shootSpeed = 0f;
